1. feature: ICP批量查询；2. feature: 爱企查；3. enhance: @xiaoliusec #16；4. enhance: @allinu #18;5. enhance: @K1p2y3 #20；6. enhance: 小程序敏感信息提取敏感信息提取规则。

+++ backend/service/model/icp/task.go
@@ -9,8 +9,8 @@ type Task struct {
 	Status       int       `json:"status"`       // 记录当前运行状态
 	ServiceTypes string    `json:"serviceTypes"` // 查询类型
 	Targets      string    `json:"targets"`      // 目标
-	Total        int       `json:"total"`        // 总数
-	Current      int       `json:"current"`      // 已查询数
+	Total        int64     `json:"total"`        // 总数
+	Current      int64     `json:"current"`      // 已查询数
 	TimeSpent    float64   `json:"timeSpent"`    // 用时s
 	Message      string    `json:"message"`
 }
